Added extensive error handling for backend operation, correct rendering of \xb1 format symbol

+++ rendering/renderClient.py
@@ -2,6 +2,8 @@ import beanstalkc
 from FuzzEd.middleware import HttpResponseServerErrorAnswer
 from django.core.mail import mail_admins
 
+wait_for = 5 #seconds
+
 def latex2eps(text):
     try:
         b=beanstalkc.Connection()
@@ -9,13 +11,17 @@ def latex2eps(text):
         b.watch('renderEpsResult')
         jobid = b.put(str(text.encode('utf-8')))
         # Wait for result ...
-        answer = b.reserve()
-        result = answer.body
-        answer.delete()
-        return result
+        answer = b.reserve(timeout=wait_for)
     except Exception as e:
         mail_admins("Error in EPS Rendering (beanstalk client)",str(e))
         raise HttpResponseServerErrorAnswer("Sorry, we have a temporary issue with the EPS rendering. Please try again later. Our admins are notified.")
+    if answer:
+        result = answer.body
+        answer.delete()
+        return result
+    else:
+        mail_admins("Error in EPS Rendering - Timeout while waiting for result",str(text))
+        raise HttpResponseServerErrorAnswer("Sorry, we have a temporary issue with the EPS rendering. Please try again later. Our admins are notified.")            
 
 def latex2pdf(text):
     try:
@@ -24,10 +30,14 @@ def latex2pdf(text):
         b.watch('renderPdfResult')
         jobid = b.put(str(text.encode('utf-8')))
         # Wait for result ...
-        answer = b.reserve()
-        result = answer.body
-        answer.delete()
-        return result
+        answer = b.reserve(timeout=wait_for)
     except Exception as e:
         mail_admins("Error in PDF Rendering (beanstalk client)",str(e))       
         raise HttpResponseServerErrorAnswer("Sorry, we have a temporary issue with the PDF rendering. Please try again later. Our admins are notified.")
+    if answer:
+        result = answer.body
+        answer.delete()
+        return result
+    else:
+        mail_admins("Error in PDF Rendering - Timeout while waiting for result",str(text))
+        raise HttpResponseServerErrorAnswer("Sorry, we have a temporary issue with the PDF rendering. Please try again later. Our admins are notified.")
